+++ app/src/main/java/br/com/guiainvestimento/fragment/BaseFragment.java
@@ -1860,7 +1860,7 @@ public abstract class BaseFragment extends Fragment {
         if(STQueryCursor.getCount() > 0){
             STQueryCursor.moveToFirst();
             // Final values to be inserted in FiiData
-            int quantityTotal = 0;
+            double quantityTotal = 0;
             double buyValue = 0;
             // Buy quantity and total is to calculate correct medium buy price
             // Medium price is only for buys
@@ -1909,14 +1909,14 @@ public abstract class BaseFragment extends Fragment {
                         buyQuantityBrokerage -= STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
                         break;
                     case Constants.Type.SPLIT:
-                        buyQuantity = buyQuantity*STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
-                        quantityTotal = quantityTotal*STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
-                        mediumPrice = mediumPrice/STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
+                        buyQuantity = buyQuantity*STQueryCursor.getDouble(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
+                        quantityTotal = quantityTotal*STQueryCursor.getDouble(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
+                        mediumPrice = mediumPrice/STQueryCursor.getDouble(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
                         break;
                     case Constants.Type.GROUPING:
-                        buyQuantity = buyQuantity/STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
-                        quantityTotal = quantityTotal/STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
-                        mediumPrice = mediumPrice*STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
+                        buyQuantity = buyQuantity/STQueryCursor.getDouble(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
+                        quantityTotal = quantityTotal/STQueryCursor.getDouble(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
+                        mediumPrice = mediumPrice*STQueryCursor.getDouble(STQueryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
                         break;
                     default:
                 }
@@ -2156,7 +2156,7 @@ public abstract class BaseFragment extends Fragment {
             do{
                 String _id = String.valueOf(queryCursor.getInt(queryCursor.getColumnIndex(PortfolioContract.FiiIncome._ID)));
                 long incomeTimestamp = queryCursor.getLong(queryCursor.getColumnIndex(PortfolioContract.FiiIncome.COLUMN_EXDIVIDEND_TIMESTAMP));
-                int quantity = getFiiQuantity(symbol, incomeTimestamp);
+                double quantity = getFiiQuantity(symbol, incomeTimestamp);
                 double perFii = queryCursor.getDouble((queryCursor.getColumnIndex(PortfolioContract.FiiIncome.COLUMN_PER_FII)));
                 int incomeType = queryCursor.getInt((queryCursor.getColumnIndex(PortfolioContract.FiiIncome.COLUMN_TYPE)));
                 double receiveTotal = quantity * perFii;
@@ -2191,7 +2191,7 @@ public abstract class BaseFragment extends Fragment {
     // Get fii quantity that will receive the dividend per fii
     // symbol is to query by specific symbol only
     // income timestamp is to query only the quantity of fiis transactions before the timestamp
-    public int getFiiQuantity(String symbol, Long incomeTimestamp){
+    public double getFiiQuantity(String symbol, Long incomeTimestamp){
         // Return column should be only quantity of fii
         String selection = PortfolioContract.FiiTransaction.COLUMN_SYMBOL + " = ? AND "
                 + PortfolioContract.FiiTransaction.COLUMN_TIMESTAMP + " < ?";
@@ -2204,7 +2204,7 @@ public abstract class BaseFragment extends Fragment {
                 null, selection, selectionArguments, sortOrder);
         if(queryCursor.getCount() > 0) {
             queryCursor.moveToFirst();
-            int quantityTotal = 0;
+            double quantityTotal = 0;
             int currentType = 0;
             do {
                 currentType = queryCursor.getInt(queryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_TYPE));
@@ -2217,10 +2217,10 @@ public abstract class BaseFragment extends Fragment {
                         quantityTotal -= queryCursor.getInt(queryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
                         break;
                     case Constants.Type.SPLIT:
-                        quantityTotal = quantityTotal*queryCursor.getInt(queryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
+                        quantityTotal = quantityTotal*queryCursor.getDouble(queryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
                         break;
                     case Constants.Type.GROUPING:
-                        quantityTotal = quantityTotal/queryCursor.getInt(queryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
+                        quantityTotal = quantityTotal/queryCursor.getDouble(queryCursor.getColumnIndex(PortfolioContract.FiiTransaction.COLUMN_QUANTITY));
                         break;
                     default:
                 }
